refactor(adapters): relocate cache/keyval/queue/secrets/schema impls
Finish F7 / Step 6 of the DDD refactor: move every concrete port
implementation out of hyperion/infrastructure/* into
hyperion/adapters/<port>/<backend>.py, and the message models into
hyperion/domain/messages.py. The old modules become deprecation shims.

- adapters/cache/{memory,filesystem,dynamodb}.py (InMemoryCache,
  LocalFileCache+DEFAULT_LOCAL_FILE_CACHE_MAX_SIZE,
  DynamoDBCache+DYNAMODB_MAX_LENGTH)
- adapters/keyval/{memory,dynamodb}.py + new lite adapters/keyval/
  filesystem.py (directory-of-files, atomic writes); is_valid_compression_type
  moved to ports/keyval.py
- adapters/queue/{memory,filesystem,sqs}.py; message models + SQS helpers
  -> domain/messages.py
- adapters/secrets/{dummy,aws_sm}.py + new lite adapters/secrets/env.py
- adapters/schema_registry/{local,s3}.py (AVRO_SCHEMAS_PATH, _get_schema_from_s3)
- adapters/http/proxy.py (redact_url, get_proxy_mounts)

Old infrastructure/* + catalog/schema.py become deprecation shims:
abstracts re-exported eagerly from ports; relocated concretes/constants
resolved lazily via importlib in __getattr__ so touching a deprecated
namespace never re-pulls boto3. PersistentCache stays in
infrastructure/cache.py (the Catalog knot removed in S7). Port from_config
deferred imports and internal importers (catalog, sources) repointed to the
new locations so the library no longer deprecation-warns itself.

No behaviour change, no new API. Every old import path resolves and emits a
DeprecationWarning. New tests: deprecation-shim identity contract, unit tests
for the two new lite adapters, import-graph guards (lite adapters boto3-free;
deferred shims boto3-free).

Closes #131.
