+++ hyperion/adapters/cache/__init__.py
@@ -1,0 +1,5 @@
+"""Cache adapters. Import the concrete module you need explicitly.
+
+``memory`` and ``filesystem`` are lite; ``dynamodb`` requires boto3 (``[aws]``).
+No eager re-exports here -- touching this namespace must stay boto3-free.
+"""
