+++ hyperion/adapters/http/__init__.py
@@ -1,0 +1,4 @@
+"""HTTP adapters. Import the concrete module you need explicitly.
+
+``proxy`` is lite (httpx only). No eager re-exports here.
+"""
