+++ hyperion/adapters/schema_registry/__init__.py
@@ -1,0 +1,5 @@
+"""Schema registry adapters. Import the concrete module you need explicitly.
+
+``local`` is lite; ``s3`` requires boto3 (``[aws]``).
+No eager re-exports here -- touching this namespace must stay boto3-free.
+"""
