+++ hyperion/ports/secrets.py
@@ -2,8 +2,9 @@
 
 Abstract :class:`SecretsManager` base and the :data:`SECRET_PATTERN` used by
 ``translate_env_vars``. Concrete adapters (``DummySecretsManager``,
-``AWSSecretsManager``) live in ``hyperion.infrastructure.secretsmanager`` until
-step S6; ``_create_new`` reaches them via a deferred import.
+``AWSSecretsManager``, ``EnvSecretsManager``) live in
+``hyperion.adapters.secrets.*``; ``_create_new`` reaches them via a deferred
+import.
 """
 
 import abc
@@ -22,8 +23,9 @@ class SecretsManager(abc.ABC):
 
     @staticmethod
     def _create_new() -> "SecretsManager":
+        from hyperion.adapters.secrets.aws_sm import AWSSecretsManager
+        from hyperion.adapters.secrets.dummy import DummySecretsManager
         from hyperion.config import secrets_config
-        from hyperion.infrastructure.secretsmanager import AWSSecretsManager, DummySecretsManager
 
         if secrets_config.backend is None:
             logger.warning("No secrets backend is configured. Using dummy secrets manager.")
